refactor the custom fields source to have a clear API for external consumption
